@@ -1,20 +1,20 @@
 module github.com/aasaam/nginx-protection
 
-go 1.17
+go 1.19
 
 require (
 	github.com/aasaam/aes-go v0.0.4
 	github.com/go-ldap/ldap/v3 v3.4.4
-	github.com/gofiber/adaptor/v2 v2.1.30
-	github.com/gofiber/fiber/v2 v2.40.1
-	github.com/gofiber/helmet/v2 v2.2.22
-	github.com/gofiber/template v1.7.3
+	github.com/gofiber/adaptor/v2 v2.1.31
+	github.com/gofiber/fiber/v2 v2.41.0
+	github.com/gofiber/helmet/v2 v2.2.23
+	github.com/gofiber/template v1.7.4
 	github.com/mdp/qrterminal v1.0.1
 	github.com/pquerna/otp v1.4.0
 	github.com/prometheus/client_golang v1.14.0
 	github.com/rs/zerolog v1.28.0
 	github.com/urfave/cli/v2 v2.23.7
-	golang.org/x/text v0.5.0
+	golang.org/x/text v0.6.0
 	gopkg.in/yaml.v2 v2.4.0
 )
 
@@ -29,7 +29,7 @@ require (
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/klauspost/compress v1.15.9 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.16 // indirect
+	github.com/mattn/go-isatty v0.0.17 // indirect
 	github.com/mattn/go-runewidth v0.0.14 // indirect
 	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
 	github.com/prometheus/client_model v0.3.0 // indirect
@@ -38,7 +38,7 @@ require (
 	github.com/rivo/uniseg v0.2.0 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasthttp v1.41.0 // indirect
+	github.com/valyala/fasthttp v1.44.0 // indirect
 	github.com/valyala/tcplisten v1.0.0 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
 	golang.org/x/crypto v0.0.0-20220622213112-05595931fe9d // indirect
